(Not working) Updated so that the board is in the game class
