+++ src/main/java/views/AcesUp/AcesUp.flt.html
@@ -111,19 +111,19 @@ console.log(game);
 
 $( '.columnOfCards .cardLocation' ).html("");
 
-    $.each(game.cols[0], function( key, val ) {
+    $.each(game.board.cols[0], function( key, val ) {
         $( '#c0 .l'+key ).html(val.value + val.suit);
     });
 
- $.each(game.cols[1], function( key, val ) {
+ $.each(game.board.cols[1], function( key, val ) {
         $( '#c1 .l'+key ).html(val.value + val.suit);
     });
 
- $.each(game.cols[2], function( key, val ) {
+ $.each(game.board.cols[2], function( key, val ) {
         $( '#c2 .l'+key ).html(val.value + val.suit);
     });
 
- $.each(game.cols[3], function( key, val ) {
+ $.each(game.board.cols[3], function( key, val ) {
         $( '#c3 .l'+key ).html(val.value + val.suit);
     });
